plugins updated

+++ wp-content/plugins/wp-fastest-cache/wpFastestCache.php
@@ -3,7 +3,7 @@
 Plugin Name: WP Fastest Cache
 Plugin URI: http://wordpress.org/plugins/wp-fastest-cache/
 Description: The simplest and fastest WP Cache system
-Version: 0.8.5.8
+Version: 0.8.6.0
 Author: Emre Vona
 Author URI: http://tr.linkedin.com/in/emrevona
 Text Domain: wp-fastest-cache
@@ -44,7 +44,6 @@ GNU General Public License for more details.
 		public $images_in_css = array("name" => "", "images" => array());
 
 		public function __construct(){
-
 			$optimize_image_ajax_requests = array("wpfc_revert_image_ajax_request", 
 												  "wpfc_statics_ajax_request",
 												  "wpfc_optimize_image_ajax_request",
@@ -136,7 +135,8 @@ GNU General Public License for more details.
 
 				$this->checkCronTime();
 
-				register_deactivation_hook( __FILE__, array('WpFastestCache', 'deactivate') );
+				register_deactivation_hook( __FILE__, array('WpFastestCache', 'deactivate'));
+				register_uninstall_hook( __FILE__, array('WpFastestCache', 'uninstall'));
 
 				if($this->isPluginActive("wp-fastest-cache-premium/wpFastestCachePremium.php")){
 					include_once $this->get_premium_path("mobile-cache.php");
@@ -193,6 +193,16 @@ GNU General Public License for more details.
 					}
 				}
 			}
+		}
+
+		public function is_trailing_slash(){
+			if($permalink_structure = get_option('permalink_structure')){
+				if(preg_match("/\/$/", $permalink_structure)){
+					return true;
+				}
+			}
+
+			return false;
 		}
 
 		public function wpfc_update_premium_callback(){
@@ -281,6 +291,12 @@ GNU General Public License for more details.
 
 		public function wpfc_check_url_ajax_request_callback(){
 			if(current_user_can('manage_options')){
+				if(preg_match("/wp\.com/", $_GET["url"])){
+					$res = array("success" => true);
+					echo json_encode($res);
+					exit;
+				}
+
 				$_GET["url"] = strip_tags($_GET["url"]);
 				$_GET["url"] = str_replace(array("'", '"'), "", $_GET["url"]);
 				
@@ -288,7 +304,7 @@ GNU General Public License for more details.
 					$_GET["url"] = "http://".$_GET["url"];
 				}
 				
-				$response = wp_remote_get($_GET["url"], array('timeout' => 10 ) );
+				$response = wp_remote_get($_GET["url"], array('timeout' => 20 ) );
 
 				$header = wp_remote_retrieve_headers($response);
 
@@ -325,14 +341,25 @@ GNU General Public License for more details.
 
 		public function wpfc_cdn_template_ajax_request_callback(){
 			if(current_user_can('manage_options')){
+				if($_POST["id"] == "maxcdn"){
+					$path = WPFC_MAIN_PATH."templates/cdn/maxcdn.php";
+				}else if($_POST["id"] == "other"){
+					$path = WPFC_MAIN_PATH."templates/cdn/other.php";
+				}else if($_POST["id"] == "photon"){
+					$path = WPFC_MAIN_PATH."templates/cdn/photon.php";
+				}else{
+					die("Wrong cdn");
+				}
+
+
 				ob_start();
-				include_once(WPFC_MAIN_PATH."templates/cdn/".$_POST["id"].".php");
+				include_once($path);
 				$content = ob_get_contents();
 				ob_end_clean();
 
 				$res = array("success" => false, "content" => "");
 
-				if($data = @file_get_contents(WPFC_MAIN_PATH."templates/cdn/".$_POST["id"].".php")){
+				if($data = @file_get_contents($path)){
 					$res["success"] = true;
 					$res["content"] = $content;
 				}
@@ -346,11 +373,27 @@ GNU General Public License for more details.
 
 		public function wpfc_save_cdn_integration_ajax_request_callback(){
 			if(current_user_can('manage_options')){
-				$values = json_encode($_POST["values"]);
-				if(get_option("WpFastestCacheCDN")){
-					update_option("WpFastestCacheCDN", $values);
+				if($data = get_option("WpFastestCacheCDN")){
+					$cdn_exist = false;
+					$arr = json_decode($data);
+
+					foreach ($arr as $cdn_key => &$cdn_value) {
+						if($cdn_value->id == $_POST["values"]["id"]){
+							$cdn_value = $_POST["values"];
+							$cdn_exist = true;
+						}
+					}
+
+					if(!$cdn_exist){
+						array_push($arr, $_POST["values"]);	
+					}
+
+					update_option("WpFastestCacheCDN", json_encode($arr));
 				}else{
-					add_option("WpFastestCacheCDN", $values, null, "yes");
+					$arr = array();
+					array_push($arr, $_POST["values"]);
+
+					add_option("WpFastestCacheCDN", json_encode($arr), null, "yes");
 				}
 				echo json_encode(array("success" => true));
 				exit;
@@ -361,7 +404,37 @@ GNU General Public License for more details.
 
 		public function wpfc_remove_cdn_integration_ajax_request_callback(){
 			if(current_user_can('manage_options')){
-				delete_option("WpFastestCacheCDN");
+    			$cdn_values = get_option("WpFastestCacheCDN");
+
+    			if($cdn_values){
+    				$std_obj = json_decode($cdn_values);
+    				$cdn_values_arr = array();
+
+    				if(is_array($std_obj)){
+						$cdn_values_arr = $std_obj;
+					}else{
+						array_push($cdn_values_arr, $std_obj);
+					}
+
+    				foreach ($cdn_values_arr as $cdn_key => $cdn_value) {
+	    				if($cdn_value->id == "amazonaws" || $cdn_value->id == "keycdn" || $cdn_value->id == "cdn77"){
+	    					$cdn_value->id = "other";
+	    				}
+
+	    				if($cdn_value->id == $_POST["id"]){
+	    					unset($cdn_values_arr[$cdn_key]);
+	    				}
+    				}
+
+    				$cdn_values_arr = array_values($cdn_values_arr);
+    			}
+
+    			if(count($cdn_values_arr) > 0){
+    				update_option("WpFastestCacheCDN", json_encode($cdn_values_arr));
+    			}else{
+					delete_option("WpFastestCacheCDN");
+    			}
+
 				echo json_encode(array("success" => true));
 				exit;
 			}else{
@@ -384,8 +457,26 @@ GNU General Public License for more details.
 		}
 
 		public function wpfc_save_exclude_pages_callback(){
+			if(!wp_verify_nonce($_POST["security"], 'wpfc-save-exclude-ajax-nonce')){
+				die( 'Security check' );
+			}
+			
 			if(current_user_can('manage_options')){
 				if(isset($_POST["rules"])){
+					foreach ($_POST["rules"] as $key => &$value) {
+						$value["prefix"] = strip_tags($value["prefix"]);
+						$value["content"] = strip_tags($value["content"]);
+
+						$value["prefix"] = preg_replace("/\=|\'|\"/", "", $value["prefix"]);
+						$value["content"] = preg_replace("/\=|\'|\"/", "", $value["content"]);
+
+						$value["content"] = trim($value["content"], "/");
+
+						if($value["prefix"] == "homepage"){
+							$this->deleteHomePageCache(false);
+						}
+					}
+
 					$data = json_encode($_POST["rules"]);
 
 					if(get_option("WpFastestCacheExclude")){
@@ -397,6 +488,8 @@ GNU General Public License for more details.
 					delete_option("WpFastestCacheExclude");
 				}
 
+				$this->modify_htaccess_for_exclude();
+
 				echo json_encode(array("success" => true));
 				exit;
 			}else{
@@ -404,7 +497,29 @@ GNU General Public License for more details.
 			}
 		}
 
+		public function modify_htaccess_for_exclude(){
+			$path = ABSPATH;
+
+			if($this->is_subdirectory_install()){
+				$path = $this->getABSPATH();
+			}
+
+			$htaccess = @file_get_contents($path.".htaccess");
+
+			if(preg_match("/\#\s?Start\sWPFC\sExclude/", $htaccess)){
+				$exclude_rules = $this->excludeRules();
+
+				$htaccess = preg_replace("/\#\s?Start\sWPFC\sExclude[^\#]*\#\s?End\sWPFC\sExclude\s+/", $exclude_rules, $htaccess);
+			}
+
+			@file_put_contents($path.".htaccess", $htaccess);
+		}
+
 		public function wpfc_save_timeout_pages_callback(){
+			if(!wp_verify_nonce($_POST["security"], 'wpfc-save-timeout-ajax-nonce')){
+				die( 'Security check' );
+			}
+
 			if(current_user_can('manage_options')){
 				$this->setCustomInterval();
 			
@@ -550,16 +665,44 @@ GNU General Public License for more details.
 				$htaccess = preg_replace("/#\s?BEGIN\s?WpFastestCache.*?#\s?END\s?WpFastestCache/s", "", $htaccess);
 				$htaccess = preg_replace("/#\s?BEGIN\s?GzipWpFastestCache.*?#\s?END\s?GzipWpFastestCache/s", "", $htaccess);
 				$htaccess = preg_replace("/#\s?BEGIN\s?LBCWpFastestCache.*?#\s?END\s?LBCWpFastestCache/s", "", $htaccess);
-				file_put_contents($path.".htaccess", $htaccess);
+				@file_put_contents($path.".htaccess", $htaccess);
 			}
 
-			wp_clear_scheduled_hook("wp_fastest_cache");
-			wp_clear_scheduled_hook($wpfc->slug()."_regular");
-			wp_clear_scheduled_hook("wp_fastest_cache_Preload");
+			$wpfc->deleteCache();
+		}
+
+		public function uninstall(){
+			$wpfc = new WpFastestCache();
+			$wpfc->deactivate();
+			
+			// wp_clear_scheduled_hook("wp_fastest_cache");
+			// wp_clear_scheduled_hook("wp_fastest_cache_regular");
 
 			delete_option("WpFastestCache");
 			delete_option("WpFcDeleteCacheLogs");
-			$wpfc->deleteCache();
+			delete_option("WpFastestCacheCDN");
+			delete_option("WpFastestCacheExclude");
+			delete_option("WpFastestCachePreLoad");
+			delete_option("WpFastestCacheCSS");
+			delete_option("WpFastestCacheCSSSIZE");
+			delete_option("WpFastestCacheJS");
+			delete_option("WpFastestCacheJSSIZE");
+
+			foreach ((array)_get_cron_array() as $cron_key => $cron_value) {
+				foreach ( (array) $cron_value as $hook => $events ) {
+					if(preg_match("/^wp\_fastest\_cache/", $hook)){
+						$args = array();
+						
+						foreach ( (array) $events as $event_key => $event ) {
+							if(isset($event["args"]) && isset($event["args"][0])){
+								$args = array(json_encode(json_decode($event["args"][0])));
+							}
+						}
+
+						wp_clear_scheduled_hook($hook, $args);
+					}
+				}
+			}
 		}
 
 		protected function slug(){
@@ -581,7 +724,7 @@ GNU General Public License for more details.
 		}
 
 		protected function get_excluded_useragent(){
-			return "facebookexternalhit";
+			return "facebookexternalhit|WhatsApp|Mediatoolkitbot";
 		}
 
 		// protected function detectNewPost(){
@@ -594,7 +737,15 @@ GNU General Public License for more details.
 			if ( ! wp_is_post_revision($post->ID) ){
 				if(isset($this->options->wpFastestCacheNewPost) && isset($this->options->wpFastestCacheStatus)){
 					if($new_status == "publish" && $old_status != "publish"){
-						$this->deleteCache();
+						if(isset($this->options->wpFastestCacheNewPost_type) && $this->options->wpFastestCacheNewPost_type){
+							if($this->options->wpFastestCacheNewPost_type == "all"){
+								$this->deleteCache();
+							}else if($this->options->wpFastestCacheNewPost_type == "homepage"){
+								$this->deleteHomePageCache();
+							}
+						}else{
+							$this->deleteCache();
+						}
 					}else if($new_status == "trash" && $old_status == "publish"){
 						$this->deleteCache();
 					}else if(($new_status == "draft" || $new_status == "pending") && $old_status == "publish"){
@@ -603,11 +754,25 @@ GNU General Public License for more details.
 				}
 
 				if($new_status == "publish" && $old_status == "publish"){
-					if(defined('WPFC_DELETE_ALL_CACHE_AFTER_UPDATE') && WPFC_DELETE_ALL_CACHE_AFTER_UPDATE){
-						$this->deleteCache();
-					}else{
-						$this->singleDeleteCache(false, $post->ID);
+					if(isset($this->options->wpFastestCacheUpdatePost) && isset($this->options->wpFastestCacheStatus)){
+
+						if($this->options->wpFastestCacheUpdatePost_type == "post"){
+							$this->singleDeleteCache(false, $post->ID);
+						}else if($this->options->wpFastestCacheUpdatePost_type == "all"){
+							$this->deleteCache();
+						}
+						
 					}
+
+
+					// if(defined('WPFC_DELETE_ALL_CACHE_AFTER_UPDATE') && WPFC_DELETE_ALL_CACHE_AFTER_UPDATE){
+					// 	$this->deleteCache();
+					// }else{
+					// 	$this->singleDeleteCache(false, $post->ID);
+					// }
+
+
+
 				}
 			}
 		}
@@ -696,6 +861,41 @@ GNU General Public License for more details.
 
 				// to check the post appears on tag
 				// toDO
+			}
+		}
+
+		public function deleteHomePageCache($log = true){
+			$site_url_path = preg_replace("/https?\:\/\/[^\/]+/i", "", site_url());
+			$home_url_path = preg_replace("/https?\:\/\/[^\/]+/i", "", home_url());
+
+			if($site_url_path){
+				$site_url_path = trim($site_url_path, "/");
+
+				if($site_url_path){
+					@unlink($this->getWpContentDir()."/cache/all/".$site_url_path."/index.html");
+					@unlink($this->getWpContentDir()."/cache/wpfc-mobile-cache/".$site_url_path."/index.html");
+				}
+			}
+
+			if($home_url_path){
+				$home_url_path = trim($home_url_path, "/");
+
+				if($home_url_path){
+					@unlink($this->getWpContentDir()."/cache/all/".$home_url_path."/index.html");
+					@unlink($this->getWpContentDir()."/cache/wpfc-mobile-cache/".$home_url_path."/index.html");
+				}
+			}
+
+			@unlink($this->getWpContentDir()."/cache/all/index.html");
+			@unlink($this->getWpContentDir()."/cache/wpfc-mobile-cache/index.html");
+
+			if($log){
+				if($this->isPluginActive("wp-fastest-cache-premium/wpFastestCachePremium.php")){
+					include_once $this->get_premium_path("logs.php");
+
+					$log = new WpFastestCacheLogs("delete");
+					$log->action();
+				}
 			}
 		}
 
@@ -1000,16 +1200,19 @@ GNU General Public License for more details.
 						$this->wpfc_remote_get(get_option("home"), "WP Fastest Cache Preload iPhone Mobile Bot - After Cache Timeout");
 					}
 				}else if($rule->prefix == "startwith"){
+						if(!is_dir($this->getWpContentDir()."/cache/tmpWpfc")){
+							if(@mkdir($this->getWpContentDir()."/cache/tmpWpfc", 0755, true)){}
+						}
 
+						$rule->content = trim($rule->content, "/");
 
+						@rename($this->getWpContentDir()."/cache/all/".$rule->content, $this->getWpContentDir()."/cache/tmpWpfc/".time());
+						@rename($this->getWpContentDir()."/cache/wpfc-mobile-cache/".$rule->content, $this->getWpContentDir()."/cache/tmpWpfc/mobile_".time());
 				}else if($rule->prefix == "exact"){
-					if(!is_dir($this->getWpContentDir()."/cache/tmpWpfc")){
-						if(@mkdir($this->getWpContentDir()."/cache/tmpWpfc", 0755, true)){}
-					}
+					$rule->content = trim($rule->content, "/");
 
-					@rename($this->getWpContentDir()."/cache/all/".$rule->content, $this->getWpContentDir()."/cache/tmpWpfc/".time());
-					@rename($this->getWpContentDir()."/cache/wpfc-mobile-cache/".$rule->content, $this->getWpContentDir()."/cache/tmpWpfc/mobile_".time());
-
+					@unlink($this->getWpContentDir()."/cache/all/".$rule->content."/index.html");
+					@unlink($this->getWpContentDir()."/cache/wpfc-mobile-cache/".$rule->content."/index.html");
 				}
 
 				if($rule->prefix != "all"){
@@ -1023,6 +1226,39 @@ GNU General Public License for more details.
 				//for old cron job
 				$this->deleteCache();
 			}
+		}
+
+		public function excludeRules(){
+			$htaccess_page_rules = "";
+			$htaccess_page_useragent = "";
+
+			if($rules_json = get_option("WpFastestCacheExclude")){
+				if($rules_json != "null"){
+					$rules_std = json_decode($rules_json);
+
+					foreach ($rules_std as $key => $value) {
+						$value->type = isset($value->type) ? $value->type : "page";
+
+						if($value->type == "page"){
+							if($value->prefix == "startwith"){
+								$htaccess_page_rules = $htaccess_page_rules."RewriteCond %{REQUEST_URI} !^/".$value->content." [NC]\n";
+							}
+
+							if($value->prefix == "contain"){
+								$htaccess_page_rules = $htaccess_page_rules."RewriteCond %{REQUEST_URI} !".$value->content." [NC]\n";
+							}
+
+							if($value->prefix == "exact"){
+								$htaccess_page_rules = $htaccess_page_rules."RewriteCond %{REQUEST_URI} !\/".$value->content." [NC]\n";
+							}
+						}else if($value->type == "useragent"){
+							$htaccess_page_useragent = $htaccess_page_useragent."RewriteCond %{HTTP_USER_AGENT} !".$value->content." [NC]\n";
+						}
+					}
+				}
+			}
+
+			return "# Start WPFC Exclude\n".$htaccess_page_rules.$htaccess_page_useragent."# End WPFC Exclude\n";
 		}
 
 		public function getABSPATH(){
@@ -1195,14 +1431,35 @@ GNU General Public License for more details.
 		}
 
 		public function cdn_replace_urls($matches){
-			$this->cdn->file_types = str_replace(",", "|", $this->cdn->file_types);
+			if(count($this->cdn) > 0){
+				foreach ($this->cdn as $key => $cdn) {
+					if(preg_match("/manifest\.json\.php/i", $matches[0])){
+						return $matches[0];
+					}
 
-			if(preg_match("/\.(".$this->cdn->file_types.")/i", $matches[0])){
-				if(preg_match("/".preg_quote($this->cdn->originurl, "/")."/", $matches[2])){
-					$matches[0] = preg_replace("/(http(s?)\:)?\/\/(www\.)?".preg_quote($this->cdn->originurl, "/")."/i", $this->cdn->cdnurl, $matches[0]);
-				}else if(preg_match("/^(\/?)(wp-includes|wp-includes)/", $matches[2])){
-					$matches[2] = preg_replace("/^\//", "", $matches[2]);
-					$matches[0] = str_replace($matches[2], $this->cdn->cdnurl."/".$matches[2], $matches[0]);
+					$cdn->file_types = str_replace(",", "|", $cdn->file_types);
+
+					if(preg_match("/\.(".$cdn->file_types.")/i", $matches[0])){
+						if($cdn->keywords){
+							$cdn->keywords = str_replace(",", "|", $cdn->keywords);
+
+							if(preg_match("/".$cdn->keywords."/i", $matches[0])){
+								if(preg_match("/".preg_quote($cdn->originurl, "/")."/", $matches[2])){
+									$matches[0] = preg_replace("/(http(s?)\:)?\/\/(www\.)?".preg_quote($cdn->originurl, "/")."/i", $cdn->cdnurl, $matches[0]);
+								}else if(preg_match("/^(\/?)(wp-includes|wp-includes)/", $matches[2])){
+									$matches[2] = preg_replace("/^\//", "", $matches[2]);
+									$matches[0] = str_replace($matches[2], $cdn->cdnurl."/".$matches[2], $matches[0]);
+								}
+							}
+						}else{
+							if(preg_match("/".preg_quote($cdn->originurl, "/")."/", $matches[2])){
+								$matches[0] = preg_replace("/(http(s?)\:)?\/\/(www\.)?".preg_quote($cdn->originurl, "/")."/i", $cdn->cdnurl, $matches[0]);
+							}else if(preg_match("/^(\/?)(wp-includes|wp-includes)/", $matches[2])){
+								$matches[2] = preg_replace("/^\//", "", $matches[2]);
+								$matches[0] = str_replace($matches[2], $cdn->cdnurl."/".$matches[2], $matches[0]);
+							}
+						}
+					}
 				}
 			}
 
@@ -1224,6 +1481,27 @@ GNU General Public License for more details.
 			}
 
 			return false;
+		}
+
+		public function set_images_in_css($css_content = false){
+			$css_content = false; // we should remove when defer css is available
+
+			if($css_content){
+				$css_content = strip_tags($css_content);
+				
+				preg_match_all("/[^\}\{]+\{[^\}\{]*url\([^\)]*\.(png|gif|jpg|jpeg)[^\)]*\)[^\}\{]*\}/i", $css_content, $out);
+
+				if(isset($out) && isset($out[0])){
+					$out[0] = preg_replace("/^\s+/", "", $out[0]);
+					$out[0] = preg_replace("/\s+/m", " ", $out[0]);
+					$out[0] = preg_replace("/\s\{\s/", "{", $out[0]);
+					$out[0] = preg_replace("/\s\}/", "}", $out[0]);
+					$out[0] = preg_replace("/[^\}\{]*(background(-image)?\s*\:\s*url\([^\(\)]+[^\;]+\;)[^\}\{]*/", "$1", $out[0]);
+
+					$GLOBALS["wp_fastest_cache"]->images_in_css["images"] = array_merge($GLOBALS["wp_fastest_cache"]->images_in_css["images"], array_unique($out[0]));
+					$GLOBALS["wp_fastest_cache"]->images_in_css["images"] = array_unique($GLOBALS["wp_fastest_cache"]->images_in_css["images"]);
+				}
+			}
 		}
 
 		public function get_operating_systems(){
